+++ src/umpire/op/CudaAdviseReadMostlyOperation.cpp
@@ -27,7 +27,7 @@ CudaAdviseReadMostlyOperation::apply(
   error = ::cudaGetDeviceProperties(&properties, 0);
 
   if (error != cudaSuccess) {
-    UMPIRE_ERROR("cudaGetDeviceProperties( device = " << device << "),"
+    UMPIRE_ERROR("cudaGetDeviceProperties( device = " << 0 << "),"
         << " failed with error: "
         << cudaGetErrorString(error));
   }